requirement.txt
streamlit
tensorflow
opencv-python
matplotlib
numpy
request